backend/pyproject.toml
[project]
name = "intentiv-backend"
version = "3.0.0"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.9.0",
    "litellm>=1.50.0",           # bridge for non-Gemini models (litellm/<provider>/<model>)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "slowapi>=0.1.9",
    # Agent Engine — required for demo/Cloud Run deployment (USE_AGENT_ENGINE=true)
    # Safe to leave installed in OSS builds; only activated when env vars are set.
    "google-cloud-aiplatform[reasoning_engines,adk]>=1.93.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "anyio>=4.0.0",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
